fix(gui): move ApiKeys helpers out of the component module
Satisfy react-refresh/only-export-components by keeping classification and
protocol helpers in api-access-models.ts.
