+++ gui/src/pages/ApiKeys.tsx
@@ -1,6 +1,12 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { IconCheck, IconPlus, IconX } from "../icons";
 import { useI18n, LOCALES } from "../i18n/shared";
+import {
+  classifyExternalModel,
+  externalModelId,
+  gatewayInboundProtocols,
+  type ExternalModelRow,
+} from "../api-access-models";
 
 interface ApiKeyEntry {
   id: string;
@@ -15,15 +21,6 @@ interface ApiEndpointInfo {
   chatCompletions: string;
   messages: string;
   models: string;
-}
-
-interface ExternalModelRow {
-  id: string;
-  displayName: string;
-  provider: string;
-  disabled?: boolean;
-  native?: boolean;
-  custom?: boolean;
 }
 
 type ModelTestState = "idle" | "testing" | "ok" | "error";
@@ -46,43 +43,6 @@ function deriveApiEndpoints(endpoint: string): ApiEndpointInfo {
     chatCompletions: `${baseUrl}/chat/completions`,
     messages: `${baseUrl}/messages`,
     models: `${baseUrl}/models`,
-  };
-}
-
-function externalModelId(model: ExternalModelRow): string {
-  return model.id;
-}
-
-/** Inbound gateway protocols — not inferred from provider type. */
-export function gatewayInboundProtocols(claudeCodeEnabled: boolean): string[] {
-  return claudeCodeEnabled
-    ? ["responses", "chat", "messages"]
-    : ["responses", "chat"];
-}
-
-/**
- * Classify a `/v1/models` row. Bare IDs keep their callable id; `owned_by`
- * decides native/combo/custom so combo aliases are not labeled OpenAI.
- */
-export function classifyExternalModel(row: {
-  id: string;
-  owned_by?: string;
-}): ExternalModelRow {
-  const slashIndex = row.id.indexOf("/");
-  const ownedBy = typeof row.owned_by === "string" && row.owned_by.trim()
-    ? row.owned_by.trim()
-    : undefined;
-  const provider = slashIndex > 0
-    ? row.id.slice(0, slashIndex)
-    : (ownedBy ?? "openai");
-  const native = slashIndex < 0 && provider === "openai";
-  const custom = provider !== "openai" && provider !== "combo";
-  return {
-    id: row.id,
-    displayName: row.id,
-    provider,
-    native,
-    custom,
   };
 }
